pyproject.toml
[project]
name = "competitor-matcher"
version = "1.0.0"
description = "Competitor Price Intelligence System for Austrian Electronics Retailers"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # HTTP/API
    "requests>=2.32.5",
    "httpx>=0.27.0",

    # Data Processing
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",

    # Vector Database (optional for CLI pipeline)
    "qdrant-client>=1.7.0",
    "fastembed>=0.2.0",

    # AI/LLM (optional)
    "anthropic>=0.20.0",

    # Web Scraping (optional)
    "playwright>=1.40.0",
    "playwright-stealth>=1.0.6",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
